# Prepare index database for multi-party submissions [KVL-700]