--- a/ledger/participant-integration-api/src/main/resources/db/migration/h2database/V28__multiple_submitters.sql
+++ b/ledger/participant-integration-api/src/main/resources/db/migration/h2database/V28__multiple_submitters.sql
@@ -1,0 +1,15 @@
+-- Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+-- SPDX-License-Identifier: Apache-2.0
+
+alter table participant_events add column submitters array;
+update participant_events set submitters = array[submitter];
+alter table participant_events drop column submitter;
+
+alter table participant_command_completions add column submitters array;
+update participant_command_completions set submitters = array[submitting_party];
+create index participant_command_completions_idx on participant_command_completions(completion_offset, application_id, submitters);
+drop index PUBLIC.INDEX_2;
+alter table participant_command_completions drop column submitting_party;
+
+-- Array subset functions
+create alias array_is_subset for "com.daml.platform.store.dao.events.SqlFunctions.arrayIsSubset"