--- a/ledger/participant-integration-api/src/main/scala/platform/store/dao/CommandCompletionsReader.scala
+++ b/ledger/participant-integration-api/src/main/scala/platform/store/dao/CommandCompletionsReader.scala
@@ -41,7 +41,7 @@ private[dao] final class CommandCompletionsReader(
       endInclusive = endInclusive,
       applicationId = applicationId,
       parties = parties,
-      sqlFunctions,
+      sqlFunctions = sqlFunctions,
     )
     Source
       .future(