--- a/ledger/participant-integration-api/src/main/scala/platform/store/dao/JdbcLedgerDao.scala
+++ b/ledger/participant-integration-api/src/main/scala/platform/store/dao/JdbcLedgerDao.scala
@@ -775,7 +775,7 @@ private class JdbcLedgerDao(
     if (submitters.length == 1)
       commandId.unwrap + "%" + submitters.head
     else
-      commandId.unwrap + "%" + submitters.asInstanceOf[List[String]].sorted.distinct.mkString("%")
+      commandId.unwrap + "%" + submitters.sorted(Ordering.String).distinct.mkString("%")
 
   override def deduplicateCommand(
       commandId: domain.CommandId,