This PR changes the index database schema such that it can support multiple submitters.

The submitters are used to filter out command IDs in the transaction stream for parties that did not submit the command, and to filter the command completion stream.