@rautenrieth-da 2020-11-27
I measured the impact of this change on the performance on the indexer. There is no evidence that this PR changes how fast the indexer can insert updates.