@rautenrieth-da 2020-12-04
Impact on the command completion stream query performance.

Index database with 200k completions, 16 different parties, 2 different applications. Querying all command completions for a given party:


### Current schema

Returning all completions within the last 100 offsets (realistic use of an app following ledger end):
```
EXPLAIN ANALYZE select completion_offset, record_time, command_id, transaction_id, status_code, status_message from participant_command_completions where completion_offset > '\x0000000000030d6a' and completion_offset <= '\x0000000000030dcd' and application_id = 'app' and submitting_party in ('1');
                                                                                                    QUERY PLAN                                                                                                     
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using participant_command_completio_completion_offset_application_idx on participant_command_completions  (cost=0.42..57.92 rows=26 width=102) (actual time=0.020..0.027 rows=2 loops=1)
   Index Cond: ((completion_offset > '\x0000000000030d6a'::bytea) AND (completion_offset <= '\x0000000000030dcd'::bytea) AND ((application_id)::text = 'app'::text) AND ((submitting_party)::text = '1'::text))
 Planning Time: 0.110 ms
 Execution Time: 0.042 ms
```

Returning all completions within the last ~30k offsets: 
```
EXPLAIN ANALYZE select completion_offset, record_time, command_id, transaction_id, status_code, status_message from participant_command_completions where completion_offset > '\x000000000002a000' and completion_offset <= '\x0000000000030dcd'  and application_id = 'app' and submitting_party in ('1');
                                                                                                       QUERY PLAN                                                                                                        
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Bitmap Heap Scan on participant_command_completions  (cost=1653.09..7925.91 rows=7465 width=102) (actual time=6.956..7.706 rows=1010 loops=1)
   Recheck Cond: ((completion_offset > '\x000000000002a000'::bytea) AND (completion_offset <= '\x0000000000030dcd'::bytea) AND ((application_id)::text = 'app'::text) AND ((submitting_party)::text = '1'::text))
   Heap Blocks: exact=373
   ->  Bitmap Index Scan on participant_command_completio_completion_offset_application_idx  (cost=0.00..1651.22 rows=7465 width=0) (actual time=6.909..6.910 rows=1010 loops=1)
         Index Cond: ((completion_offset > '\x000000000002a000'::bytea) AND (completion_offset <= '\x0000000000030dcd'::bytea) AND ((application_id)::text = 'app'::text) AND ((submitting_party)::text = '1'::text))
 Planning Time: 0.097 ms
 Execution Time: 7.796 ms
(7 rows)
```

### New schema

Returning all completions within the last 100 offsets (realistic use of an app following ledger end):
```
EXPLAIN ANALYZE select completion_offset, record_time, command_id, transaction_id, status_code, status_message from participant_command_completions where completion_offset > '\x0000000000030d6a' and completion_offset <= '\x0000000000030dcd' and application_id = 'app' and submitting_party && array['1']::varchar[];
                                                                                             QUERY PLAN                                                                                             
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using participant_command_completio_completion_offset_application_idx on participant_command_completions  (cost=0.42..94.27 rows=1 width=102) (actual time=0.026..0.065 rows=2 loops=1)
   Index Cond: ((completion_offset > '\x0000000000030d6a'::bytea) AND (completion_offset <= '\x0000000000030dcd'::bytea) AND ((application_id)::text = 'app'::text))
   Filter: (submitting_party && '{1}'::character varying[])
   Rows Removed by Filter: 50
 Planning Time: 0.114 ms
 Execution Time: 0.079 ms
(6 rows)
```

Returning all completions within the last ~30k offsets: 
```
EXPLAIN ANALYZE select completion_offset, record_time, command_id, transaction_id, status_code, status_message from participant_command_completions where completion_offset > '\x000000000002a000' and completion_offset <= '\x0000000000030dcd' and application_id = 'app' and submitting_party && array['1']::varchar[];
                                                                                    QUERY PLAN                                                                                     
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Bitmap Heap Scan on participant_command_completions  (cost=1337.44..5177.74 rows=70 width=102) (actual time=4.614..7.858 rows=1010 loops=1)
   Recheck Cond: ((completion_offset > '\x000000000002a000'::bytea) AND (completion_offset <= '\x0000000000030dcd'::bytea) AND ((application_id)::text = 'app'::text))
   Filter: (submitting_party && '{1}'::character varying[])
   Rows Removed by Filter: 13047
   Heap Blocks: exact=511
   ->  Bitmap Index Scan on participant_command_completio_completion_offset_application_idx  (cost=0.00..1337.42 rows=14015 width=0) (actual time=4.556..4.556 rows=14057 loops=1)
         Index Cond: ((completion_offset > '\x000000000002a000'::bytea) AND (completion_offset <= '\x0000000000030dcd'::bytea) AND ((application_id)::text = 'app'::text))
 Planning Time: 0.110 ms
 Execution Time: 7.931 ms
(9 rows)

```